Matrix_extensions compilable on Windows

+++ dwtools/Matrix_extensions.h
@@ -50,13 +50,13 @@ void Matrix_scale (Matrix me, int choice);
 autoMatrix Matrix_transpose (Matrix me);
 
 void Matrix_drawDistribution (Matrix me, Graphics g, double xmin, double xmax,
-	double ymin, double ymax, double minimum, double maximum, long nBins,
+	double ymin, double ymax, double minimum, double maximum, integer nBins,
 	double freqMin, double freqmax, bool cumulative, bool garnish);
 
 void Matrix_drawSliceY (Matrix me, Graphics g, double x, double ymin, double ymax,
 	double min, double max);
 
-int Matrix_fitPolynomial (Matrix me, long maxDegree);
+int Matrix_fitPolynomial (Matrix me, integer maxDegree);
 
 autoMatrix Matrix_solveEquation (Matrix me, double tolerance);
